Refactor Implemented endpoints properly Assigned variables as needed

+++ Nova/src/app/components/navbar/navbar.component.ts
@@ -7,6 +7,8 @@ import { Product } from 'src/app/interfaces/product';
 import { DataService } from 'src/app/services/data.service';
 import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
+import {AuthService} from "../../services/auth.service";
+import {CurrentUser} from "../../classes/user";
 
 
 @Component({
@@ -21,13 +23,10 @@ export class NavbarComponent implements OnInit, OnDestroy {
   productNames: String[] = [];
   productsService: ProductsService;
   searchForm!: FormGroup;
-
-  username: String |null = sessionStorage.getItem("username");
-
+  username: String | undefined = CurrentUser.username;
   message!: String;
   sent!: String;
   subscription!: Subscription;
-//   search: String = '';
   navbarOpen = false;
 
   ngOnInit(): void {
@@ -53,7 +52,7 @@ export class NavbarComponent implements OnInit, OnDestroy {
       this.showDropDown = false;
     }, 100)
     //console.log('...timeout passed.');
-    
+
   }
 
   //Check If the value of the form group exists or not
@@ -65,7 +64,11 @@ export class NavbarComponent implements OnInit, OnDestroy {
     }
   }
 
-  constructor( private fb: FormBuilder, _productsService: ProductsService, private data: DataService, private router: Router) {
+  constructor( private fb: FormBuilder,
+               private _productsService: ProductsService,
+               private data: DataService,
+               private router: Router,
+               private auth: AuthService){
     this.initForm()
     this.productsService = _productsService;
   }
@@ -87,11 +90,11 @@ export class NavbarComponent implements OnInit, OnDestroy {
     let a = this.searchForm.value.search;
     this.data.changeMessage(a);
     this.data.changeSent('true');
-    
+
   }
 
   selectValue(value: any) {
-    
+
     this.searchForm.patchValue({"search": value});
     this.showDropDown = false;
     this.searchFor({'search': value})
@@ -111,13 +114,17 @@ export class NavbarComponent implements OnInit, OnDestroy {
   toggleNavbar() {
     this.navbarOpen = !this.navbarOpen
     console.log("clicked")
-  }  
+  }
 
   logout(){
-    sessionStorage.clear();
-    // alert("You are now logged out")
-    console.log("logged out")
-    // this.router.navigate(['/'])
+    this.auth.logout()
+      .subscribe(resp => {
+        if (resp.body == 'Successful Logout')
+        {
+          sessionStorage.clear();
+          this.router.navigate(['/']);
+        }
+      })
   }
 }
